fix: working around apparent bug in Jenkins' Last Changes plugin.
Since we moved the sandbox directories to galileo, we've been seeing
issues similar to those described at
https://github.com/jenkinsci/last-changes-plugin/issues/49. Our case
appears to have something to do with NFS, and having Last Changes use
"PREVIOUS_REVISION" instead of "LAST_SUCCESSFUL_BUILD" as described in
the bug report seems to work around the issue. Oddly, this doesn't seem
to happen on every project, but for now making that change allows
Jenkins to build proejcts again.