Error handling improvements to Architect and Keymaster
Added 'components_not_in_state' function to the Architect, which gives
an easy way to find out which devices are not yet in a desired state.
Added error handling to the heartbeat thread, so that it doesn't kill
everything when it cannot talk to the keymaster server for some reason.